+++ ui/protection_paladin/inputs.ts
@@ -57,6 +57,11 @@ export const PrimarySealSelection = InputHelpers.makeSpecOptionsEnumIconInput<Sp
 			actionId: () => ActionId.fromSpellId(407798),
 			value: PaladinSeal.Martyrdom,
 		},
+		{
+			actionId: () => ActionId.fromSpellId(20423),
+			value: PaladinSeal.Fury,
+			showWhen: (player: Player<Spec.SpecProtectionPaladin>) => player.getTalents().sealOfFury,
+		},
 	],
 	changeEmitter: (player: Player<Spec.SpecProtectionPaladin>) =>
 		TypedEvent.onAny([player.gearChangeEmitter, player.talentsChangeEmitter, player.specOptionsChangeEmitter]),